+++ djpcms/utils/__init__.py
@@ -28,24 +28,6 @@ def isexact(bit):
     else:
         return bit
 
-def form_kwargs(request, instance = None, withrequest = False, **kwargs):
-    '''
-    Quick form arguments aggregator
-    
-    Usage:
-      def someview(request):
-          form = MyForm(**form_kwargs(request))
-    '''
-    if request and request.method == 'POST':
-        kwargs['data'] = request.POST
-        kwargs['files'] = request.FILES
-    if withrequest:
-        kwargs['request'] = request
-    if instance:
-        kwargs['instance'] = instance
-    return kwargs
-
-
 
 class NoAjaxKeyError(Exception):
     
@@ -152,13 +134,12 @@ class requestwrap(UnicodeObject):
 def form_kwargs(request,
                 instance = None, withrequest = False,
                 withdata = True, method = 'POST', **kwargs):
-    '''
-    Quick form arguments aggregator
-    
-    Usage:
-      def someview(request):
-          form = MyForm(**form_kwargs(request))
-    '''
+    '''Quick form arguments aggregator.
+Usage::
+
+    form = MyForm(**form_kwargs(request))
+
+'''
     if request and withdata and request.method == method:
         kwargs['data'] = request.POST
         kwargs['files'] = request.FILES