+++ djpcms/utils/navigation.py
@@ -1,6 +1,7 @@
 '''
 Utility module for creating a navigation list
 '''
+from djpcms.template import loader
 from djpcms.utils import lazyattr    
 
 
@@ -11,6 +12,7 @@ class lazycounter(object):
     def __new__(cls, djp, **kwargs):
         obj = super(lazycounter, cls).__new__(cls)
         obj.djp     = djp
+        obj.classes = kwargs.pop('classes',None)
         obj.kwargs  = kwargs
         return obj
 
@@ -45,11 +47,21 @@ class Navigator(lazycounter):
     '''
     A navigator for a web site
     '''
+    def __init__(self, *args, **kwargs):
+        self.name    = self.kwargs.pop('name','')
+        self.url     = self.kwargs.pop('url','')
+        self.levels  = self.kwargs.pop('levels',1)
+        self.mylevel = self.kwargs.pop('mylevel',1)
+        self.liclass = self.kwargs.pop('liclass',None)
+        
     def make_item(self, djp, classes):
-        return {'name':    djp.linkname,
-                'url':     djp.url,
-                'classes': u' '.join(classes),
-                'nav':     Navigator(djp, **self.kwargs)}
+        return Navigator(djp,
+                         levels = self.levels,
+                         mylevel = self.mylevel+1,
+                         liclass = classes,
+                         url = djp.url,
+                         name = djp.linkname,
+                         **self.kwargs)
     
     def buildselects(self, djp, urlselects):
         parent = djp.parent
@@ -84,9 +96,14 @@ class Navigator(lazycounter):
                 classes.append(scn)
             if url in urlselects:
                 classes.append(css.link_selected)
-            items.append(self.make_item(djp, classes))
+            items.append(self.make_item(djp, u' '.join(classes)))
         return items
 
+    def render(self):
+        if self.mylevel <= self.levels:
+            return loader.render_to_string('djpcms/bits/navitem.html', {'navigator': self})
+        else:
+            return u''         
 
 
 class Breadcrumbs(lazycounter):