+++ docs/source/conf.py
@@ -24,7 +24,7 @@ sys.path.insert(0,base_dir)
 # The short X.Y version.
 version = '0.7'
 # The full version, including alpha/beta/rc tags.
-release = '0.7.3'
+release = '0.7.4'
 
 # -- General configuration -----------------------------------------------------